refactor menu functions to methods
